+++ CHANGELOG.md
@@ -18,6 +18,12 @@ Backend
 - Minimum JAX version bumped up to ``0.4.29``
 
 
+Performance Improvements
+
+- [Partial summation in coordinate mapping](https://github.com/PlasmaControl/DESC/pull/1826).
+- [NUFFTS](https://github.com/PlasmaControl/DESC/pull/1834) are now used by default for computing bounce integrals.
+
+
 v0.15.0
 -------